fix: updated ConmpanyPlayerMatchStats with repository and service

+++ src/main/java/us/lsi/dp1/newcorporder/stats/companyPlayerMatchStats/CompanyPlayerMatchStatsService.java
@@ -27,7 +27,7 @@ public class CompanyPlayerMatchStatsService {
         return statsRepository.findByCompanyTypeAndAbilityUsedGreaterThan(companyType, abilityUsedThreshold);
     }
 
-    public Optional<CompanyPlayerMatchStats> getStatsById(Long id) {return statsRepository.findById(id);}
+    public Optional<CompanyPlayerMatchStats> getStatsById(Integer id) {return statsRepository.findById(id);}
 
     public List<CompanyPlayerMatchStats> getStatsOrderedByAbilityUsedDesc() {return statsRepository.findByOrderByAbilityUsedDesc();}
